// src/pages/ArticlePage/model/types/articlesPageSheme.ts
import { Article, ArticleSortField, ArticleType, ArticleView } from '@/entities/Article';
import { SortOrder } from '@/shared/types/sort';
import { EntityState } from '@reduxjs/toolkit';

export interface ArticlesPageScheme extends EntityState<Article> {
	isLoading?: boolean;
	error?: string;
	// pagination
	page: number;
	limit: number;
	hasMore: boolean;

	// filters
	view: ArticleView;
	order: SortOrder;
	sort: ArticleSortField;
	search: string;
	type: ArticleType;
	_inited: boolean;
}